requirements.txt
atomicwrites==1.3.0
attrs==22.2.0
colorama==0.4.1
enum34==1.1.6
lxml==4.9.2
more-itertools==7.0.0
namedlist==1.7
pluggy==1.0.0
py==1.8.0
pytest==7.2.0
pytest-allure-adaptor2==1.8.2
six==1.12.0
requests==2.28.1
